netrender: moved transport specific code to netrender_transport.*
